Added `de/tuple_structs` module
